backend/requirements.txt
Flask
flask-cors
langchain-google-genai
langchain
langsmith
python-dotenv
